refactor: move should_skip_path to utils/path_utils.py
